Replace unsigned ints with plain ints
As it is generally advised for arithmetic integer expressions to use
signed ints instead of their unsigned counterparts, all occurences of
unsigned ints have been replaced. Now it is a tad easier to check for
errors. The usage of ints raises the possibility of overflows and
therefore undefined behaviour. The CMake file enables UBSanitizers as
countermeasure.